+++ packages/likert-choice/src/main.jsx
@@ -1,23 +1,8 @@
 import React from 'react';
-import ChoiceInput from './choice-input';
-import {withStyles} from 'material-ui/styles';
-
-const styleSheet = {
-    container : {
-        display: 'table',
-        width: '100%',
-        tableLayout: 'fixed',
-        textAlign: 'center',
-        borderSpacing: '10px'
-    }
-}
+import LikertChoice from './likert-choice';
 
 const Main = (props) => {
-    return (
-        <div className={props.classes.container}>
-            <ChoiceInput displayKey="1" checked={true} disabled={false} label="Hello" onChange={() => console.log("hello")} value="rost" />
-        </div>
-    )
+    return <LikertChoice {...props}/>
 }
 
-export default withStyles(styleSheet)(Main);
+export default Main;
